+++ Engine/Src/Gear/Core/DX/ReadbackHeap.cpp
@@ -1,6 +1,6 @@
 #include<Gear/Core/DX/ReadbackHeap.h>
 
-ReadbackHeap::ReadbackHeap(const UINT size, const D3D12_HEAP_FLAGS flags) :
+ReadbackHeap::ReadbackHeap(const uint64_t size, const D3D12_HEAP_FLAGS flags) :
 	Resource(CD3DX12_HEAP_PROPERTIES(D3D12_HEAP_TYPE_READBACK), flags, CD3DX12_RESOURCE_DESC::Buffer(size), false, D3D12_RESOURCE_STATE_COPY_DEST, nullptr)
 {
 }
